+++ src/modules/filters/index.ts
@@ -3,6 +3,10 @@ import VoxelGrid from './VoxelGrid';
 import StatisticalOutlierRemoval from './StatisticalOutlierRemoval';
 import RadiusOutlierRemoval from './RadiusOutlierRemoval';
 import UniformSampling from './UniformSampling';
+import RandomSample from './RandomSample';
+import GridMinimum from './GridMinimum';
+import LocalMaximum from './LocalMaximum';
+import ApproximateVoxelGrid from './ApproximateVoxelGrid';
 
 export default {
   PassThrough,
@@ -10,4 +14,8 @@ export default {
   StatisticalOutlierRemoval,
   RadiusOutlierRemoval,
   UniformSampling,
+  RandomSample,
+  GridMinimum,
+  LocalMaximum,
+  ApproximateVoxelGrid,
 };