+++ tests/common.ts
@@ -2,16 +2,17 @@
 // @ts-nocheck
 import fs from 'fs';
 import path from 'path';
-import { performance } from 'perf_hooks';
 import * as PCL from '../';
 
-global.performance = performance;
-
-const wasm = fs.readFileSync(path.join(__dirname, '../dist/pcl-core.wasm'));
-
 export function initPCL() {
+  const wasm = fs.readFileSync(path.join(__dirname, '../dist/pcl-core.wasm'));
   return PCL.init({
     arrayBuffer: wasm,
     log: false,
   });
+}
+
+export function writeFile(name: string, pcl: PCL.PCLInstance) {
+  const pcd = fs.readFileSync(path.join(__dirname, `../data/${name}`));
+  pcl.fs.writeFile(name, new Uint8Array(pcd));
 }